Add the functions of failure recovery and auto-scaling. For auto-scaling, there are three types: 1. Sub-topology level horizonal scaling up or down. For this type, the entire requested sub-topology can be scaled up or down.
2. VM level horizonal scaling.
For this type, the requested VM can be scaled up. These VM can come from different sub-topologies and the network topology is scaled according to the original VM at the same time.

3. VM level vertical scaling.
For this type, the requested VM can be vertically scaled up or down. Vertical scaling here means that the computing capability of the VM can be scaled. For instance, original VM "nodeA" has 1 vCPU and 1GB memory. According to the request, it can be scaled to 1 vCPU and 3GB memory. Meanwhile, the network connection remains the same.

I also design the infrastructure code for achieving this.

For sub-topology level horizonal scaling, the infrastructure code is simillar as:

- CodeType: "SEQ"
  OpCode:
    Operation: "hscale"
    Options:
        ReqID: "hst1"
        CP: "ExoGENI"
        DC: "GWU (Washington DC,  USA) XO Rack"
        UpDown: UP
        ScaledSTName: null
        STClass: null
        SEClass: null
        VEClass: null
    ObjectType: SubTopology
    Objects: "scaling_1||scaling_2"
- CodeType: "SEQ"
  OpCode:
    Operation: "hscale"
    ObjectType: REQ
    Objects: "hst1"

For VM level horizonal scaling, the infrastructure code is simillar as:

- CodeType: "SEQ"
  OpCode:
    Operation: "hscale"
    Options:
        ReqID: "hvm1"
        CP: "ExoGENI"
        DC: "GWU (Washington DC,  USA) XO Rack"
        ScaledSTName: null
        STClass: null
        SEClass: null
        VEClass: null
    ObjectType: VM
    Objects: "scaling_1.Node0||scaling_2.Node2"
- CodeType: "SEQ"
  OpCode:
    Operation: "hscale"
    ObjectType: REQ
    Objects: "hvm1"

For VM level vertical scaling, the infrastructure code is simillar as:

- CodeType: "SEQ"
  OpCode:
    Operation: "vscale"
    Options:
        ReqID: "v1"
        CPU: 1
        MEM: 3
        STClass: null
        SEClass: null
        VEClass: null
    ObjectType: VM
    Objects: "scaling_1.Node0||scaling_2.Node2"
- CodeType: "SEQ"
  OpCode:
    Operation: "vscale"
    ObjectType: REQ
    Objects: "v1"

All the field with the value of "null" can be omitted.
Currently, all these functions have been test with ExoGENI.

+++ src/main/java/provisioning/engine/SEngine/EC2SEngine.java
@@ -1,8 +1,11 @@
 package provisioning.engine.SEngine;
 
+import java.io.File;
 import java.util.UUID;
 
 import org.apache.log4j.Logger;
+
+import commonTool.CommonTool;
 
 import provisioning.credential.Credential;
 import provisioning.credential.SSHKeyPair;
@@ -64,22 +67,41 @@ public class EC2SEngine extends SEngine {
 	public boolean createAccessSSHKey(SubTopologyInfo subTopologyInfo, 
 			Credential credential, Database database) {
 		SubTopology ec2SubTopology = subTopologyInfo.subTopology;
+		String cp = subTopologyInfo.cloudProvider.trim().toLowerCase();
+		String dc = subTopologyInfo.domain.trim().toLowerCase();
 		//create a key pair for this sub-topology, if there is none.
 		if(ec2SubTopology.accessKeyPair == null){
-			String keyPairId = UUID.randomUUID().toString();
-			String publicKeyId = "publicKey-"+keyPairId;
-			String privateKeyString = 
-					EC2VEngine.createSSHKeyPair(subTopologyInfo, 
-							credential, publicKeyId);
-			if(privateKeyString == null){
-				logger.error("Unexpected error for creating ssh key pair for sub-topology '"+ec2SubTopology.topologyName+"'!");
-				return false;
-			}
+			String keyPairId = cp+"-"+dc;
 			subTopologyInfo.sshKeyPairId = keyPairId;
-			ec2SubTopology.accessKeyPair = new SSHKeyPair();
-			ec2SubTopology.accessKeyPair.publicKeyId = publicKeyId;
-			ec2SubTopology.accessKeyPair.privateKeyString = privateKeyString;
-			ec2SubTopology.accessKeyPair.SSHKeyPairId = keyPairId;
+			String currentDir = CommonTool.getPathDir(ec2SubTopology.loadingPath);
+            String sshKeyDir = currentDir + keyPairId + File.separator;
+            File keyDir = new File(sshKeyDir);
+            if (!keyDir.exists()) {
+                logger.info("There is no ssh key pair for sub-topology '" + ec2SubTopology.topologyName + "'! Generating!");
+                
+                String publicKeyId = "publicKey-"+UUID.randomUUID().toString();
+	    			String privateKeyString = 
+	    					EC2VEngine.createSSHKeyPair(subTopologyInfo, 
+	    							credential, publicKeyId);
+	    			if(privateKeyString == null){
+	    				logger.error("Unexpected error for creating ssh key pair for sub-topology '"+ec2SubTopology.topologyName+"'!");
+	    				return false;
+	    			}
+	    			ec2SubTopology.accessKeyPair = new SSHKeyPair();
+	    			ec2SubTopology.accessKeyPair.publicKeyId = publicKeyId;
+	    			ec2SubTopology.accessKeyPair.privateKeyString = privateKeyString;
+	    			ec2SubTopology.accessKeyPair.SSHKeyPairId = keyPairId;
+                
+            } else {
+                logger.info("The ssh key pair for sub-topology '" + ec2SubTopology.topologyName + "' has already exist!");
+                ec2SubTopology.accessKeyPair = new SSHKeyPair();
+                if(!ec2SubTopology.accessKeyPair.loadSSHKeyPair(keyPairId, sshKeyDir)){
+                		logger.error("Error when loading SSH key Pair from "+sshKeyDir);
+                		return false;
+                }
+            }
+
+			
 		}
 		return true;
 	}
@@ -112,6 +134,9 @@ public class EC2SEngine extends SEngine {
 		////leverage VEngine to delete all the VMs
 		if(!super.delete(subTopologyInfo, credential, database))
 			return false;
+		
+		if(subTopologyInfo.topology.startsWith("_tmp_"))
+			return true;
 		
 		if(!EC2VEngine.deleteVPC(subTopologyInfo, credential))
 			return false;
